refactor family list
