--- src/components/families/family-list/FamilyList.tsx
@@ -1,51 +1,0 @@
-import React, { useState } from "react";
-
-import FamilyAPI from "api/FamilyAPI";
-import { FamilyDetailResponse, FamilyListResponse } from "api/types";
-import { DefaultField } from "types";
-
-import FamilySidebar from "./family-sidebar";
-import FamilyTable from "./family-table";
-
-type Props = {
-  families: FamilyListResponse[];
-  enrolmentFields: DefaultField[];
-  shouldDisplayDynamicFields: boolean;
-};
-
-const FamilyList = ({
-  families,
-  enrolmentFields,
-  shouldDisplayDynamicFields,
-}: Props) => {
-  const [selectedFamily, setSelectedFamily] = useState<FamilyDetailResponse>();
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
-
-  const onSelectFamily = async (id: number) => {
-    const family = await FamilyAPI.getFamilyById(id);
-    setSelectedFamily(family);
-    setIsSidebarOpen(true);
-  };
-
-  const handleCloseSidebar = () => setIsSidebarOpen(false);
-
-  return (
-    <>
-      <FamilyTable
-        families={families}
-        enrolmentFields={enrolmentFields}
-        shouldDisplayDynamicFields={shouldDisplayDynamicFields}
-        onSelectFamily={onSelectFamily}
-      />
-      {selectedFamily && (
-        <FamilySidebar
-          isOpen={isSidebarOpen}
-          family={selectedFamily}
-          onClose={handleCloseSidebar}
-        />
-      )}
-    </>
-  );
-};
-
-export default FamilyList;
