--- src/components/families/family-list/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FamilyList";
