+++ src/pages/Sessions.tsx
@@ -13,14 +13,17 @@ import { Add } from "@material-ui/icons/";
 import { useHistory, useParams } from "react-router-dom";
 
 import ClassAPI from "api/ClassAPI";
+import FamilyAPI from "api/FamilyAPI";
 import SessionAPI from "api/SessionAPI";
 import {
   ClassDetailResponse,
   SessionListResponse,
   SessionDetailResponse,
   FamilyListResponse,
+  FamilyDetailResponse,
 } from "api/types";
-import FamilyList from "components/families/family-list";
+import FamilySidebar from "components/families/family-sidebar";
+import FamilyTable from "components/families/family-table";
 import RegistrationDialog from "components/registration/RegistrationDialog";
 import SessionDetailView, {
   ALL_CLASSES_TAB_INDEX,
@@ -51,6 +54,11 @@ const Sessions = () => {
   const [classTabIndex, setClassTabIndex] = useState(ALL_CLASSES_TAB_INDEX);
   const [isLoadingSession, setIsLoadingSession] = useState(true);
   const [displayRegDialog, setDisplayRegDialog] = useState(false);
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+  const [
+    selectedFamily,
+    setSelectedFamily,
+  ] = useState<FamilyDetailResponse | null>(null);
 
   const updateSelectedSession = async (id: number) => {
     await setSelectedSession(await SessionAPI.getSession(id));
@@ -137,6 +145,12 @@ const Sessions = () => {
     ? [DefaultFields.CURRENT_CLASS]
     : [];
 
+  const onSelectFamily = async (id: number) => {
+    const family = await FamilyAPI.getFamilyById(id);
+    setSelectedFamily(family);
+    setIsSidebarOpen(true);
+  };
+
   return (
     <>
       <Box display="flex">
@@ -188,11 +202,21 @@ const Sessions = () => {
           classTabIndex={classTabIndex}
           onChangeClassTabIndex={handleChangeClassTabIndex}
           classDefaultView={
-            <FamilyList
-              families={getFamilies()}
-              enrolmentFields={getEnrolmentFields}
-              shouldDisplayDynamicFields={false}
-            />
+            <>
+              <FamilyTable
+                families={getFamilies()}
+                enrolmentFields={getEnrolmentFields}
+                shouldDisplayDynamicFields={false}
+                onSelectFamily={onSelectFamily}
+              />
+              {selectedFamily && (
+                <FamilySidebar
+                  isOpen={isSidebarOpen}
+                  family={selectedFamily}
+                  onClose={() => setIsSidebarOpen(false)}
+                />
+              )}
+            </>
           }
         />
       )}
